Add support for multiple clusters w/ same name in different namespaces
Previously, the ArgoCD cluster secret was named by the cluster itself. This could've caused conflicts
if there were clusters in other namespaces with the same name.

To address this, the controller will now generate secret names with the format <NAMESPACE>-<NAME>, to ensure uniqueness.

Some extra changes in logging and additional tests are added too.